SIH Backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.117.1
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
